+++ demisto_sdk/commands/common/hook_validations/deprecation.py
@@ -10,37 +10,33 @@ class DeprecationValidator:
     def __init__(self, id_set_file: Dict[str, List]):
         self.script_section = id_set_file.get("scripts", [])
         self.playbook_section = id_set_file.get("playbooks", [])
-        self.test_playbook_section = id_set_file.get("TestPlaybooks", [])
 
-    def validate_integartion(self, deprecated_commands_list: List[str], test_playbooks_ls: List[str]):
+    def validate_integartion(self, deprecated_commands_list: List[str]):
         """
         Manages the deprecation usage for integration commands
-        Checks if the given deprecated integration commands are used in a none-deprecated scripts / playbooks / testplaybooks
+        Checks if the given deprecated integration commands are used in a none-deprecated scripts / playbooks
 
         Args:
             deprecated_commands_list (list): A list of all the integration's deprecated commands.
-            test_playbooks_ls (list): A list of all the tests that're related to the given integration.
 
         Return:
-            dict: A dictionary where the keys are the integartion's deprecated commands that are used in none-deprecated  scripts / playbooks / testplaybooks.
+            dict: A dictionary where the keys are the integartion's deprecated commands that are used in none-deprecated  scripts / playbooks.
             The values are the file names where they're being used
         """
         usage_dict: Dict[str, list] = {}
 
-        self.filter_testplaybooks_for_integration_validation(deprecated_commands_list, test_playbooks_ls, usage_dict)
         self.filter_playbooks_for_integration_validation(deprecated_commands_list, usage_dict)
         self.find_scripts_using_given_integration_commands(deprecated_commands_list, usage_dict)
 
         return usage_dict
 
-    def validate_playbook(self, playbook_name: str, test_playbooks_ls: List[str]):
+    def validate_playbook(self, playbook_name: str):
         """
         Manages the deprecation usage validation for playbooks.
-        Checks if the given deprecated playbook is used in a none-deprecated playbooks / testplaybooks.
+        Checks if the given deprecated playbook is used in a none-deprecated playbooks.
 
         Args:
             playbook_name (str): The name of the playbook.
-            test_playbooks_ls (list): A list of all the tests that're related to the given playbook.
 
         Return:
             list: A list of all the none-deprecated files that are using the given playbook.
@@ -48,19 +44,17 @@ class DeprecationValidator:
         usage_list: List[str] = []
         key_to_check = "implementing_playbooks"
 
-        self.filter_testplaybooks_for_scripts_or_playbook_validation(test_playbooks_ls, playbook_name, usage_list, key_to_check)
         self.filter_playbooks_for_scripts_or_playbook_validation(playbook_name, usage_list, key_to_check)
 
         return usage_list
 
-    def validate_script(self, script_name: str, test_playbooks_ls: List[str]):
+    def validate_script(self, script_name: str):
         """
         Manages the deprecation usage validation for scripts.
-        Checks if the given deprecated script is used in a none-deprecated playbooks / testplaybooks / scripts.
+        Checks if the given deprecated script is used in a none-deprecated playbooks / scripts.
 
         Args:
             script_name (str): The name of the script.
-            test_playbooks_ls (list): A list of all the tests that're related to the given playbook.
 
         Return:
             list: A list of all the none-deprecated files that are using the given script.
@@ -69,33 +63,9 @@ class DeprecationValidator:
         key_to_check = "implementing_scripts"
 
         self.find_scripts_using_given_script(script_name, usage_list)
-        self.filter_testplaybooks_for_scripts_or_playbook_validation(test_playbooks_ls, script_name, usage_list, key_to_check)
         self.filter_playbooks_for_scripts_or_playbook_validation(script_name, usage_list, key_to_check)
 
         return usage_list
-
-    def filter_testplaybooks_for_integration_validation(self, deprecated_commands_list: List[str], test_playbooks_ls: List[str], usage_dict: Dict[str, list]):
-        """
-        Filter the relevant test_playbooks for the current integration validation from the test_playbook_section
-        and check which of the integration commands are being used in this files using the validate_integration_not_in_playbook function.
-
-        Args:
-            deprecated_commands_list (list): A list of all the integration's deprecated commands.
-            test_playbooks_ls (list): A list of all the tests that're related to the given integration.
-            usage_dict (dict): A dictionary where the keys are the integartion's deprecated commands
-            that are used in none-deprecated scripts / playbooks / testplaybooks.
-            The values are the file names where they're being used.
-        """
-        if test_playbooks_ls:
-            for test_playbook in self.test_playbook_section:
-                for test_playbook_val in test_playbook.values():
-                    if test_playbook_val.get("name", "") in test_playbooks_ls:
-                        command_to_integration = test_playbook_val.get("command_to_integration")
-                        if command_to_integration:
-                            self.validate_integration_commands_not_in_playbook(usage_dict, deprecated_commands_list, command_to_integration, test_playbook_val)
-                        test_playbooks_ls.remove(test_playbook_val.get("name", ""))
-                        if not test_playbooks_ls:
-                            return
 
     def filter_playbooks_for_integration_validation(self, deprecated_commands_list, usage_dict: Dict[str, list]):
         """
@@ -105,7 +75,7 @@ class DeprecationValidator:
         Args:
             deprecated_commands_list (list): A list of all the integration's deprecated commands.
             usage_dict (dict): A dictionary where the keys are the integartion's deprecated commands
-            that are used in none-deprecated scripts / playbooks / testplaybooks.
+            that are used in none-deprecated scripts / playbooks.
             The values are the file names where they're being used.
         """
         for playbook in self.playbook_section:
@@ -123,7 +93,7 @@ class DeprecationValidator:
         Args:
             deprecated_commands_list (list): A list of all the integration's deprecated commands.
             usage_dict (dict): A dictionary where the keys are the integartion's deprecated commands
-            that are used in none-deprecated scripts / playbooks / testplaybooks.
+            that are used in none-deprecated scripts / playbooks.
             The values are the file names where they're being used.
             command_to_integration(dict) A dict where the keys are the integartions commands that are being used
             and the value is the integration name.
@@ -147,7 +117,7 @@ class DeprecationValidator:
         Args:
             deprecated_commands_list (list): A list of all the integration's deprecated commands.
             usage_dict (dict): A dictionary where the keys are the integartion's deprecated commands
-            that are used in none-deprecated scripts / playbooks / testplaybooks.
+            that are used in none-deprecated scripts / playbooks.
             The values are the file names where they're being used.
         """
         for script in self.script_section:
@@ -179,28 +149,6 @@ class DeprecationValidator:
                 depends_commads_list = script_val.get("depends_on")
                 if depends_commads_list and script_name in depends_commads_list:
                     usage_list.append(script_val.get("file_path"))
-
-    def filter_testplaybooks_for_scripts_or_playbook_validation(self, test_playbooks_ls: List[str], curent_entity_name: str,
-                                                                usage_list: List[str], key_to_check: str):
-        """
-        Filter the relevant testplaybooks for the current script / playbook validation from the test_playbook_section.
-
-        Args:
-            test_playbooks_ls (list): A list of all the tests that're related to the given integration.
-            curent_entity_name (str): The name of the script / playbook currently being checked.
-            usage_list (list): A list of all the file paths that are using the script / playbook currently being checked.
-            key_to_check (str): The field in which the right entity to check is located at.
-        """
-        if test_playbooks_ls:
-            for test_playbook in self.test_playbook_section:
-                for test_playbook_val in test_playbook.values():
-                    if test_playbook_val.get("name", "") in test_playbooks_ls:
-                        implementing_entities = test_playbook_val.get(key_to_check)
-                        if implementing_entities:
-                            self.validate_playbook_or_script_not_in_playbook(usage_list, curent_entity_name, implementing_entities, test_playbook_val)
-                        test_playbooks_ls.remove(test_playbook_val.get("name", ""))
-                        if not test_playbooks_ls:
-                            return
 
     def filter_playbooks_for_scripts_or_playbook_validation(self, curent_entity_name: str, usage_list: List[str], key_to_check: str):
         """