+++ demisto_sdk/commands/common/tools.py
@@ -2789,23 +2789,6 @@ def remove_copy_and_dev_suffixes_from_str(field_name: str) -> str:
     return field_name
 
 
-def extract_testplaybooks_list(yml_file):
-    """
-    Get a yml file and extract that tests playbook section from it
-
-    Args:
-        yml_file (dict): The content of the given yml file.
-
-    Returns:
-        list[str]: A list with the names of the testplaybooks that are listed in the tests sections of the given yml
-                   or an empty list if there's no tests section / tests section include "No Tests"
-    """
-    if tests_section := yml_file.get("tests"):
-        if "No tests" not in tests_section[0]:
-            return tests_section
-    return []
-
-
 def get_invalid_incident_fields_from_mapper(
     mapper_incident_fields: Dict[str, Dict], mapping_type: str, content_fields: List
 ) -> List[str]: