+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Fixed an issue where YAML files contained incorrect value type for `tests` key when running `format --deprecate`.
 * Added a deprecation message to the `tests:` section of yaml files when running `format --deprecate`.
 * Added the 'integration-get-indicators' commands to be ignored by the **verify_yml_commands_match_readme** validation, the validation will no longer fail if these commands are not in the readme file.
-* Added ***deprecation*** a validation class that lists all the file paths of files that are using a deprecated integration command / script / playbook.
+* Added **validate** check that list all the file paths of files that are using a deprecated integration command / script / playbook.
 * Added a new validation to the **validate** command to verify that if the phrase "breaking changes" is present in a pack release notes, a JSON file with the same name exists and contains the relevant breaking changes information.
 
 ## 1.6.6